requirements.txt
Faker==37.1.0
psycopg2-binary==2.9.10
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
shiboken6==6.6.1
tzdata==2025.2
